backend/requirements.txt
aniso8601==9.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-RESTful==0.3.9
geographiclib==1.52
geopy==2.2.0
google-api-core==2.5.0
google-auth==1.35.0
google-cloud-dialogflow==2.11.0
googleapis-common-protos==1.54.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
